Hi everyone! We have an important update to share today.
As the popularity and number of Discord bots grow, it's important to keep our users and developers safe and healthy. This means from time to time, like any maturing  platform, we need to update our policies to reflect the current needs of the ecosystem.

In April of 2022, message content will become a Privileged Intent. Like other privileged intents (Presence and Guild Members), it will require approval and affect only verified bots.

If you run a verified bot, or a bot that you plan to get verified in the future, we recommend starting to transition to new functionality like slash commands and other interactions.

We want to help make this transition as easy as possible, so we're starting off by providing plenty of time (nine months) as well as hosting an "office hours" where you can ask us questions or flag any considerations. This will be on Wednesday, August 4, at 11am pacific. We will be sharing our answers to questions that come in for those who want to read but can't make that time.

We're also continuing to invest in interactions as the future of bots, including new slash command option types, better moderation controls, and new types of interactions.

- You can read the full details in our help article here, along with an FAQ: https://support-dev.discord.com/hc/en-us/articles/4404772028055
- If you have questions or concerns, we'd appreciate you filling out this very short survey to help us centralize feedback: https://s.alchemer.com/s3/f1811257ea63. We'll use these responses as discussion topics for part of the office hours and to help inform our roadmap.
- To keep this server organized, please keep discussion of this change in intents . Discussion in other channels will be removed to ensure that your fellow developers with other questions about the API can get answers.

Thank you all for your help in making this change with us. We know it's a big one, and we know that the future of bots on Discord remains incredibly exciting.